Just kidding, guess I forgot to commit. Here's the changes.
Alright, so:

Functionality is there, both in scatterplot and threed_bar_chart.

Scatterplot is finished, minus commenting code.

Threed_bar_chart, I need to go back and remove chunks of code that
are extraneous or commented out, and then comment it myself. I also
need to fix the labels displaying kind of crooked. Other than that,
though, it is finished: all the functionality is there.